requirements.txt
asgiref==3.8.1
Django==4.2
django-autoslug==1.9.9
pillow==10.3.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
tzdata==2024.1
